18/06/2015

+++ EC-Admin/EC-Admin/Forms/Compra/frmNuevaCompra.cs
@@ -47,7 +47,7 @@ namespace EC_Admin.Forms
             foreach (DataGridViewRow dr in dgvProductos.Rows)
             {
                 c.IDProductos.Add((int)dr.Cells[0].Value);
-                c.Cantidad.Add((decimal)dr.Cells[4].Value);
+                c.Cantidad.Add((int)dr.Cells[4].Value);
                 c.Unidad.Add((Unidades)Enum.Parse(typeof(Unidades), dr.Cells[6].Value.ToString()));
                 c.Precio.Add((decimal)dr.Cells[3].Value);
                 c.DescuentoProducto.Add((decimal)dr.Cells[5].Value);
