+++ EC-Admin/EC-Admin/Forms/Promocion/frmEditarPromocion.cs
@@ -64,18 +64,18 @@ namespace EC_Admin.Forms
             {
                 p.IDProducto = id;
                 p.Existencias = chbExistencias.Checked;
-                p.Cantidad = decimal.Parse(txtCant.Text);
+                p.Cantidad = int.Parse(txtCant.Text);
                 if (chbExistencias.Checked)
                 {
                     p.FechaInicio = new DateTime();
                     p.FechaFin = new DateTime();
-                    p.CantidadProducto = decimal.Parse(txtExistencias.Text);
+                    p.CantidadProducto = int.Parse(txtExistencias.Text);
                 }
                 else
                 {
                     p.FechaInicio = dtpFechaIni.Value;
                     p.FechaFin = dtpFechaFin.Value;
-                    p.CantidadProducto = 0M;
+                    p.CantidadProducto = 0;
                 }
                 p.Precio = decimal.Parse(txtPrecio.Text);
                 p.Editar();
@@ -124,7 +124,7 @@ namespace EC_Admin.Forms
             }
             else
             {
-                if (decimal.Parse(txtCant.Text) <= 0)
+                if (int.Parse(txtCant.Text) <= 0)
                 {
                     FuncionesGenerales.ColoresError(txtCant);
                     res = false;
@@ -168,7 +168,7 @@ namespace EC_Admin.Forms
                 foreach (DataRow dr in dt.Rows)
                 {
                     lblProducto.Text = dr["nombre"].ToString();
-                    cant = (decimal)dr["cant"];
+                    cant = (int)dr["cant"];
                 }
             }
             catch (MySqlException ex)
